pwnable.kr/pyproject.toml
[project]
name = "pwnable-kr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.6.0",
    "pwntools>=4.15.0",
]
